Module ply_rs::ply

Definitions used to model PLY files.

Structs

ConsistencyError
Contains a description, why a given Ply object isn't consistent and could not be made consistent.

ElementDef
Models the definition of an element.

Header
Models the header of a PLY file.

Ply
Models all necessary information to interact with a PLY file.

PropertyDef
Defines a property of an element.

Version
Models a version number.

Enums

Encoding
Models possible encoding standards for the payload.

Property
Wrapper used to implement a dynamic type system as required by the PLY file format.

PropertyType
Data type used to encode properties in the payload.

ScalarType
Scalar type used to encode properties in the payload.

Traits

Addable
Convenience trait to assure consistency between map key and name attribute of stored element.

Key
Convenience trait to assure consistency between the key used for storage and the name of the elment.

PropertyAccess
Provides setters and getters for the Parser and the Writer.

Type Definitions

Comment
Alias to give comments an explicit type.

DefaultElement
Ready to use data-structure for all kind of element definitions.

KeyMap
Alias to reduce coupling with LinkedHashMap

ObjInfo
Alias to give object informations an explicit type.

Payload
The part after end_header, contains the main data.