Struct ply_rs::ply::ElementDef

pub struct ElementDef {
    pub name: String,
    pub count: usize,
    pub properties: KeyMap<PropertyDef>,
}

Models the definition of an element.

Elements describe single entities consisting of different properties. A single point is an element. We might model it as consisting of three coordinates: x, y, and z. Usually, one finds a list of elements in a ply file.

Fields
name: String

Name of the element.

Each element within a PLY file needs a unique name. There are common conventions like using "vertex" and "face" to assure interoperability between applications. For further information, please consult your target applications or the original specification.

count: usize

Describes, how many elements appear in a PLY file.

The count is used when reading since we need to know how many elements we should interprete as having this type. The count is also needed for writing, since it will be written to the header.

properties: KeyMap<PropertyDef>

An element is modeled by multiple properties, those are named values or lists.

Examples

Point: We can define a point by its three coordinates. Hence we have three properties: x, y, and z. Reasonable types would be float or double.

Polygon: A polygon can be defined as a list of points. Since the points are stored in a list, we can define a list of indices. Good types would be some of the unsigned integer lists.

Methods

impl ElementDef

fn new(name: String) -> Self

Creates a new element definition.

The name should be unique for each element in a PLY file.

You should never need to set count manuall, since it is set by the consistency check (see make_consistent() of Ply).

No properties are set.