Struct ply_rs::ply::Ply

pub struct Ply<E: PropertyAccess> {
    pub header: Header,
    pub payload: Payload<E>,
}

Models all necessary information to interact with a PLY file.

The generic parameter E is the element type used to store the payload data.

Fields
header: Header

All header information found in a PLY file.

payload: Payload<E>

The payloud found after the end_header line in a PLY file.

One line in an ascii PLY file corresponds to a single element. The payload groups elments with the same type together in a vector.

Examples

Assume you have a Ply object called ply and want to access the third point element:

// get ply from somewhere ...
let ref a_point = ply.payload["point"][2];
let ref a_point_x = ply.payload["point"][2]["x"];

Methods

impl<E: PropertyAccess> Ply<E>

fn make_consistent(&mut self) -> Result<(), ConsistencyError>

Takes a mutable Ply object, performs commen operations to make it consistent,

When written, a consistent Ply object generates a valid PLY file. This method also checks for invariants that can't be fixed automatically. If something can not be fixed automatically, it returns a ConsistencyError describing the problem.

Remarks

This method should always be called before writing to a file with Writer. Only exception is write_ply(), which, for convenience, performs the check itself. See write_ply_unchecked() for a variant that expects the client to assure consistency.

impl<E: PropertyAccess> Ply<E>

fn new() -> Self

Creates a new Ply<E>.